refactor(core): split out instance and route through workspaces
